refactor: extract WeekPanel into a separate file
